+++ README.md
@@ -12,40 +12,37 @@ Adds additional rendering support to markdown in JupyterLab by using [markdown-i
 
 ## Prerequisites
 
-- JupyterLab
+- JupyterLab 2
 
 ## Installation
-
-Ensure mimetype exists for wasm:
-
-```bash
-echo "application/wasm      wasm" | sudo tee -a /etc/mime.types
-```
 
 Install extension:
 
 ```bash
+pip install jupyterlab_markup
 jupyter labextension install @agoose77/jupyterlab-markup
 ```
 
-You will need to reinstall the extension if you do these steps out of order.
-
-## Development
-
-For a development install (requires npm version 4 or later), do the following in the repository directory:
+Ensure the extensions are enabled:
 
 ```bash
-npm install
-npm run build
-jupyter labextension link .
+jupyter labextension list      # should contain @goose/jupyterlab-markup
+jupyter serverextension list   # should contain jupyterlab_markup
 ```
 
-To rebuild the package and the JupyterLab app:
+If the serverextension is missing, try...
 
 ```bash
-npm run build
-jupyter lab build
+jupyter serverextension enable jupyterlab_markup
 ```
+
+## Contributing
+
+Please see the [contributor guide](./CONTRIBUTING.md)!
+
+## Examples
+
+### mermaid
 
 ```mermaid
 graph TD;
@@ -54,6 +51,8 @@ graph TD;
     B-->D;
     C-->D;
 ```
+
+### svgbob
 
 ```bob
      .---.
