+++ package.json
@@ -53,6 +53,6 @@
     "typescript": "~3.7.3"
   },
   "jupyterlab": {
-    "extension": true
+    "extension": "lib/plugin.js"
   }
 }
